Separate AudioService from main
